requirements.txt
# ID Immobilier — Dépendances Python

# Data manipulation
pandas==2.2.0
numpy==1.26.4
openpyxl==3.1.2

# Big Data
pyspark==3.5.1

# Base de données MongoDB
pymongo==4.11.1
motor==3.6.1

# API
fastapi==0.115.6
uvicorn==0.32.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0

# Orchestration
apache-airflow==2.8.2

# Dashboard & Visualisation
streamlit==1.32.0
plotly==5.19.0
folium==0.16.0
streamlit-folium==0.19.0

# Utilitaires
python-dotenv==1.0.1
jupyter==1.0.0
ipykernel==6.29.3
